+++ src/wchar/mb_cur_max.c
@@ -1,6 +1,6 @@
 #include <stdlib.h>
 
-int __mb_cur_max(void)
+size_t __mb_cur_max(void)
 {
     return MB_CUR_MAX;
 }
